+++ docs/changelog.rst
@@ -7,6 +7,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 0.15.2 [future]
 --------------------
 
+- hints kitten: Allow pressing :sc:`goto_file_line` to quickly open
+  the selected file at the selected line in vim (:iss:`2268`)
+
 - Allow choosing OpenType features for individual fonts via the
   :opt:`font_features` option.